+++ source/SAMRAI/hier/PatchHierarchy.h
@@ -933,7 +933,7 @@ public:
       const char *note,
       int ln,
       int cycle,
-      int level_time,
+      double level_time,
       bool log_fine_connector,
       bool log_coarse_connector) const;
